requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

httpx==0.27.2
pydantic-settings==2.6.1
python-multipart==0.0.12

yt-dlp==2025.1.26

pillow==10.4.0
qrcode==8.0

# Optional berat (tapi kita bikin endpoint-nya graceful)
rembg==2.0.61
onnxruntime==1.20.1gemini-webapi
httpx==0.28.1
moviepy==2.2.1
pdf2docx==0.5.8
python-multipart==0.0.12
